jjb/mkck: Split make-check nightly template into sles and opensuse
For SLES and openSUSE are different build steps needed.
On SLES, artifacts needs to be syncronized and the artifacts
repositories need to be setup on the target jenkins slave.
On openSUSE, there are currently no artifacts to sync. Instead, adding
the default repositories is needed to have needed packages available.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>