jjb/mkck: Fix artifacts and repo setup for nightly runs
There are different requirements for openSUSE Leap and SLES.
On SLES, artifacts need to be delivered and used on the dynamicaly
created Jenkins slave.
On OpenSUSE, standard repositories can be used and no artifacts are
needed.
To meet both requirements, use JJB macros and make add some noop
macros which can be used for openSUSE.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>